refactor: use tokio::sync::RwLock
In the case of futures_locks::RwLock, readers and writers
are queued independently. There is no way for the scheduler
to interleave lock acquisitions with this model. An async
version of the test shows the expected
readers-starve-writers behavior.

tokio::sync::RwLock uses a waiter queue where
readers acquire 1 permit from a semaphore, and
writers acquire all permits. The waiter queue
interleaves read and write lock acquisitions in FIFO order,
so it does provide a task-fair locking policy! Hooray!

https://www.reddit.com/r/rust/comments/f4zldz/i_audited_3_different_implementation_of_async/